requirements.txt
pyyaml>=6.0
colorama>=0.4.6
pytest>=7.4.0
pytest-cov>=4.1.0
